+++ packages/core/tests/api/AsyncStateManager/SPECIAL-init-no-repeat.test.ts
@@ -1,4 +1,5 @@
-import { CleanupManager, Nullable, TestUtils } from '../../test-helpers'
+/* eslint-disable no-console */
+import { CleanupManager, Nullable } from '../../test-helpers'
 import { TestConfig, wrapper } from '../../test-wrapper'
 
 wrapper(({ buildEnv, Lib: { AsyncStateManager } }: TestConfig) => {
@@ -7,7 +8,6 @@ wrapper(({ buildEnv, Lib: { AsyncStateManager } }: TestConfig) => {
 
   const cleanupManager = new CleanupManager()
   afterEach(cleanupManager.performCleanup)
-  TestUtils.spyOnConsoleError()
 
   test('commit', async () => {
     const TestState = new AsyncStateManager(0)
@@ -21,7 +21,6 @@ wrapper(({ buildEnv, Lib: { AsyncStateManager } }: TestConfig) => {
     cleanupManager.append(TestState.isInitializing.watch(watchCallback))
     commitRef(2)
     expect(watchCallback).not.toHaveBeenCalled()
-    // eslint-disable-next-line no-console
     expect(console.error).toHaveBeenCalledTimes(buildEnv === 'prod' ? 0 : 1)
   })
 
@@ -37,7 +36,6 @@ wrapper(({ buildEnv, Lib: { AsyncStateManager } }: TestConfig) => {
     cleanupManager.append(TestState.isInitializing.watch(watchCallback))
     commitRef()
     expect(watchCallback).not.toHaveBeenCalled()
-    // eslint-disable-next-line no-console
     expect(console.error).toHaveBeenCalledTimes(buildEnv === 'prod' ? 0 : 1)
   })