+++ packages/core/tests/api/StateManager/SPECIAL-init-vs-set.test.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import { CleanupManager, TestUtils } from '../../test-helpers'
 import { TestConfig, wrapper } from '../../test-wrapper'
 
@@ -7,7 +8,6 @@ wrapper(({ buildEnv, Lib: { StateManager } }: TestConfig) => {
 
   const cleanupManager = new CleanupManager()
   afterEach(cleanupManager.performCleanup)
-  TestUtils.spyOnConsoleError()
 
   test('commit', async () => {
 
@@ -19,7 +19,6 @@ wrapper(({ buildEnv, Lib: { StateManager } }: TestConfig) => {
     })
 
     TestState.set(41)
-    // eslint-disable-next-line no-console
     expect(console.error).toHaveBeenCalledTimes(buildEnv === 'prod' ? 0 : 1)
     expect(TestState.get()).toBe(0)
 
@@ -41,7 +40,6 @@ wrapper(({ buildEnv, Lib: { StateManager } }: TestConfig) => {
     })
 
     TestState.set(41)
-    // eslint-disable-next-line no-console
     expect(console.error).toHaveBeenCalledTimes(buildEnv === 'prod' ? 0 : 1)
     expect(TestState.get()).toBe(0)