+++ packages/react/jest.setup.ts
@@ -25,23 +25,3 @@ afterEach(() => {
     console[consoleMethodName] = originalMethods[consoleMethodName] as any
   }
 })
-
-// TODO: [Low priority]
-// https://jestjs.io/docs/expect#expectextendmatchers
-// https://stackoverflow.com/a/64471550/5810737
-// import type { TestConfig } from './tests/test-wrapper'
-// expect.extend({
-//   toHaveBeenCalledIfNotProdEnv: (received, buildEnv: TestConfig['buildEnv']) => {
-//     if (buildEnv !== 'prod') {
-//       return {
-//         pass: true,
-//         message: () => `Expected \`console.error\` to be called but it is not`,
-//       }
-//     } else {
-//       return {
-//         pass: received,
-//         message: () => '',
-//       }
-//     }
-//   }
-// })