+++ packages/core/CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changelog
 
-* Initial release
+* Fixed a bug where the `didReset` lifecycle hook is called during initialization via `commit`.