+++ packages/react/CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changelog
 
-* Initial release
+* Passing `StateManager` and `AsyncStateManager` to `useSimpleStateValue` and `useSimpleStateValueWithReactiveSelector` will now cause an error to be thrown.